Add --trust-policy option to unpack command
