+++ Manifest.txt
@@ -389,6 +389,7 @@ lib/rubygems/security/policies.rb
 lib/rubygems/security/policy.rb
 lib/rubygems/security/signer.rb
 lib/rubygems/security/trust_dir.rb
+lib/rubygems/security_option.rb
 lib/rubygems/server.rb
 lib/rubygems/source.rb
 lib/rubygems/source/git.rb
